+++ eo-runtime/src/main/java/EOorg/EOeolang/EOsm/Win32/RecvFuncCall.java
@@ -38,7 +38,7 @@ public final class RecvFuncCall implements Syscall {
     public Phi make(final Phi... params) {
         final Phi result = this.win.take("return").copy();
         final int size = new Dataized(params[1]).asNumber().intValue();
-        final byte[] buf = new byte[(int) size];
+        final byte[] buf = new byte[size];
         final int received = Winsock.INSTANCE.recv(
             new Dataized(params[0]).asNumber().intValue(),
             buf,
